+++ src/burn/engine/precomp.h
@@ -85,7 +85,7 @@
 #include "detect.h"
 #include "plan.h"
 #include "logging.h"
-#include "pipe.h"
+#include "burnpipe.h"
 #include "cache.h"
 #include "dependency.h"
 #include "core.h"
